Adding GBK encoding (fixing issue 7) Recode tables are using word instead of int (prssrenc.dll is 2 times smaller now - yay!)

+++ prssrenc/encodings/cp1252.cpp
@@ -28,7 +28,7 @@
 
 #include "../codepages.h"
 
-DLLEXT int cpCP1252[256] = {
+DLLEXT WORD cpCP1252[256] = {
 	0x0000,
 	0x0001,
 	0x0002,
